chore: Convert float to double, allocate embeddings on heap

+++ include/emb.h
@@ -1,5 +1,5 @@
 #ifndef mintdb_mini_emb_h
 #define mintdb_mini_emb_h
-float cosine_distance(float a[], float b[], int size);
+double cosine_distance(double a[], double b[], int size);
 
 #endif
